fix registration flow

+++ app/src/main/java/com/example/vin/metron/entities/PDAMRecord.kt
@@ -6,6 +6,6 @@ data class PDAMRecord(
     var no_pdam: String?,
     var time_start: Timestamp?,
     var time_end: Timestamp?,
-    var number_read: Float?,
-    var usage: Float?
+    var number_read: Double?,
+    var usage: Double?
 )
