+++ app/src/main/java/com/example/vin/metron/entities/PLNRecord.kt
@@ -6,6 +6,6 @@ data class PLNRecord(
     var no_pln: String?,
     var time_start: Timestamp?,
     var time_end: Timestamp?,
-    var number_read: Float?,
-    var usage: Float?
+    var number_read: Double?,
+    var usage: Double?
 )
